chaos: fix lint

+++ src/main/java/net/kuina/nebulaecraft/creativetab/TabNebulaecraftColor.java
@@ -22,7 +22,7 @@ public class TabNebulaecraftColor extends ElementsNebulaecraftMod.ModElement {
 			@SideOnly(Side.CLIENT)
 			@Override
 			public ItemStack getTabIconItem() {
-				return new ItemStack(BlockColor1.block, (int) (1));
+				return new ItemStack(BlockColor1.block, 1);
 			}
 
 			@SideOnly(Side.CLIENT)
