+++ src/main/java/net/kuina/nebulaecraft/creativetab/TabNebulaecraftRoad.java
@@ -22,7 +22,7 @@ public class TabNebulaecraftRoad extends ElementsNebulaecraftMod.ModElement {
 			@SideOnly(Side.CLIENT)
 			@Override
 			public ItemStack getTabIconItem() {
-				return new ItemStack(BlockAsphalt.block, (int) (1));
+				return new ItemStack(BlockAsphalt.block, 1);
 			}
 
 			@SideOnly(Side.CLIENT)
